+++ source/extensions/isaacsim.replicator.writers/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [1.3.0] - 2026-02-03
+### Changed
+- Added pinholeOpenCV and fisheyePolynomial projection support to pose writer
+- Moved DOPE utils to DOPEWriter class
+ 
 ## [1.2.1] - 2025-12-01
 ### Changed
 - Deprecate DOPEWriter and YCBVideoWriter writers